requirement.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
fastapi==0.95.0
fonttools==4.38.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.3.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.21.6
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
pip==22.3.1
prettytable==3.7.0
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
setuptools==65.6.3
shyaml==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
torch==1.13.1
torchaudio==0.13.1
torchinfo==1.7.2
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.1
wheel==0.38.4
zipp==3.15.0
